style: MVC framework of backend
