+++ api/booking/booking.py
@@ -1,7 +1,5 @@
-from flask import Blueprint, render_template, session, request, redirect, url_for, jsonify
-from modules import connect_to_db
-from modules.create_jwt import decode_token
-from modules.connect_to_db import conn
+from flask import Blueprint, render_template, request, jsonify
+from .models import *
 
 booking = Blueprint(
     "booking", __name__,
@@ -9,141 +7,65 @@ booking = Blueprint(
     template_folder = "templates"
     )
 
+@booking.route("/booking")
+def base():
+	return render_template("booking.html")
+
 @booking.route("/api/booking", methods = ["GET", "POST", "DELETE"])
 def api_booking():
+#------------get booking data controller-------------
     if request.method == "GET":
-        cookiesToken = request.cookies.get("token")
-        if cookiesToken == None:
+        data = get_booking_data()
+        if data == False:
             return jsonify({
-                        "error": True,
-                        "message": "未登入系統"
-                        }), 403
-        else:
-            try:
-                decodeToken = decode_token(cookiesToken)
-                c = conn()
-                cur = c.cursor(dictionary=True)
-                sql = '''
-                SELECT 
-                    user.name,
-                    user.email,
-                    attraction.id, 
-                    attraction.name, 
-                    attraction.address, 
-                    attraction.images, 
-                    booking.date, 
-                    booking.time, 
-                    booking.price,
-                    booking.booking_id
-                FROM 
-                    booking 
-                INNER JOIN 
-                    attraction 
-                ON 
-                    booking.attractionId = attraction.id
-                INNER JOIN 
-                    user 
-                ON 
-                    booking.user_id = user.id
-                WHERE
-                    user.id = %s
-                '''
-                cur.execute(sql,(decodeToken["id"],))
-                query = cur.fetchall()
-                if query:
-                    i = 0
-                    list = []
-                    while i < len(query):
-                        result = {
-                            "attraction": {
-                            "id": query[i]["id"],
-                            "name": query[i]["name"],
-                            "address": query[i]["address"],
-                            "image": eval(query[i]["images"])[0]
-                            },
-                        "date": query[i]["date"],
-                        "time": query[i]["time"],
-                        "price": query[i]["price"],
-                        "booking_id": query[i]["booking_id"]
-                        } 
-                        list.append(result)
-                        i = i + 1
-                    return jsonify({
-                        "data": list
-                    }), 200
-                else:
-                    return jsonify({
-                        "data": None
-                        }), 200
-            finally:
-                c.close()
+                "error": True,
+                "message": "未登入系統"
+                }), 403
 
-                    
+        if data:
+            return jsonify({
+                "data": data
+                }), 200
+
+        if data == None:
+            return jsonify({
+                "data": None
+                }), 200
+#------------post booking data controller-------------
     if request.method == "POST":
-        try:
-            data = request.get_json()
-            cookiesToken = request.cookies.get("token")
-            decodeToken = decode_token(cookiesToken)
-            c = conn()
-            if data["attractionId"] == None or data["date"] == None or data["time"] == None or data["price"] == None:
-                return jsonify({
-                        "error": True,
-                        "message": "建立失敗，請確認是否欄位都有填寫"
-                        }), 400
-            elif cookiesToken == None:
-                return jsonify({
-                        "error": True,
-                        "message": "未登入系統"
-                        }), 403
-            else:
-                cur = c.cursor()
-                sql = '''
-                INSERT INTO 
-                    booking (
-                        user_id, 
-                        attractionId, 
-                        date, 
-                        time, 
-                        price
-                        ) 
-                    VALUES (%s, %s, %s, %s, %s)'''
-                cur.execute(sql, (decodeToken["id"], data["attractionId"], data["date"], data["time"], data["price"],))
-                c.commit()
-                cur.close()
-                return jsonify({
-                        "ok": True
-                        }), 200
-        except:
+        data = post_booking_data()
+        if data == "建立失敗":
+             return jsonify({
+                "error": True,
+                "message": "建立失敗，請確認是否欄位都有填寫"
+                }), 400
+
+        if data == False:
+            return jsonify({
+                "error": True,
+                "message": "未登入系統"
+                }), 403
+
+        if data == True:
+            return jsonify({
+                "data": None
+                }), 200
+
+        if data == "內部伺服器錯誤":
             return jsonify({
                 "error": True,
                 "message": "內部伺服器錯誤"
                 }), 500
-        finally:
-            c.close()
-
+#------------delete booking data controller-------------
     if request.method == "DELETE":
-        cookiesToken = request.cookies.get("token")
-        if cookiesToken == None:
+        data = delete_booking_data()
+        if data == False:
             return jsonify({
-                        "error": True,
-                        "message": "未登入系統"
-                        }), 403
-        else:
-            try:
-                data = request.get_json()
-                c = conn()
-                cur = c.cursor()
-                sql = "DELETE FROM booking WHERE booking_id = %s"
-                cur.execute(sql, (data["booking_id"],))
-                c.commit()
-                cur.close()
-                return jsonify({
-                            "ok": True
-                            }), 200
-            finally:
-                c.close()
+                "error": True,
+                "message": "未登入系統"
+                }), 403
 
-
-@booking.route("/booking")
-def base():
-	return render_template("booking.html")
+        if data == True:
+            return jsonify({
+                    "ok": True
+                    }), 200
