+++ app.py
@@ -21,9 +21,5 @@ app.register_blueprint(thankyou)
 def index():
 	return render_template("index.html")
     
-@app.route("/attraction/<id>")
-def attraction(id):
-	return render_template("attraction.html")
-
 if __name__ == "__main__":
     app.run(host="0.0.0.0", port=3000, debug=True)
